Add CloneWindowRole property
The highlight overlay is on top of everything on the desktop. This new
property allow to the user to "show" a window above the overlay.

This new property is a window role that will be cloned and drawn on top
of the highlight actor.

This cloned actor is not reactive, so it will not possible to interact
with the cloned window that's on top of the highlight overlay.

https://phabricator.endlessm.com/T30672